wargame/UIT_CTF/Super_safe_token/src/app/requirements.txt
flask
mysql-connector
pyjwt
requests
cryptography